serverUtils.js --> .ts
